Code for list of employee Sorting based on input
There is an employee class with multiple properties like firstName,lastName,Age and Country . Design program which will take input as a list of employees with given properties and the second argument is an array of properties i.e. first name, last name ,age and country combination as dynamic string array input and employee list should be sorted based on array.